chore: update log file with integration test run entries

+++ Backend/src/main/java/com/thesystem/modules/xp/dto/XpAccountResponse.java
@@ -12,7 +12,7 @@ public record XpAccountResponse(
         Integer totalXpEarned,
         Integer totalXpSpent,
         Integer lifetimeXp,
-        Integer levelProgress,
+        Double levelProgress,
         Instant createdAt,
         Instant updatedAt
 ) {
