+++ Backend/src/main/java/com/thesystem/modules/xp/dto/statistics/StatisticsResponse.java
@@ -11,7 +11,7 @@ public record StatisticsResponse(
         Integer monthlyXp,
         Integer lifetimeXp,
         Integer currentLevel,
-        Integer levelProgress,
+        Double levelProgress,
         Integer tasksCompleted,
         Integer goalsCompleted,
         Integer achievementsUnlocked,
